Remove PontoConnect module
