+++ lib/ibanity/api/ponto_connect/synchronization.ex
@@ -150,7 +150,7 @@ defmodule Ibanity.PontoConnect.Synchronization do
 
   def find(other, _id) do
     raise ArgumentError,
-      message: PontoConnect.token_argument_error_msg("Synchronization", other)
+      message: PontoConnect.RequestUtils.token_argument_error_msg("Synchronization", other)
   end
 
   @doc false
