+++ lib/ibanity/api/ponto_connect/transaction.ex
@@ -88,7 +88,7 @@ defmodule Ibanity.PontoConnect.Transaction do
 
   def list(other, _account_id) do
     raise ArgumentError,
-      message: PontoConnect.token_argument_error_msg("Transactions", other)
+      message: PontoConnect.RequestUtils.token_argument_error_msg("Transactions", other)
   end
 
   @doc """
@@ -156,7 +156,7 @@ defmodule Ibanity.PontoConnect.Transaction do
 
   def list_updated_for_synchronization(other, _synchronization_id) do
     raise ArgumentError,
-      message: PontoConnect.token_argument_error_msg("Transactions", other)
+      message: PontoConnect.RequestUtils.token_argument_error_msg("Transactions", other)
   end
 
   @doc """
@@ -222,7 +222,7 @@ defmodule Ibanity.PontoConnect.Transaction do
 
   def list_updated_pending_for_synchronization(other, _synchronization_id) do
     raise ArgumentError,
-      message: PontoConnect.token_argument_error_msg("Transactions", other)
+      message: PontoConnect.RequestUtils.token_argument_error_msg("Transactions", other)
   end
 
   @doc """
@@ -281,7 +281,7 @@ defmodule Ibanity.PontoConnect.Transaction do
 
   def list_pending_for_account(other, _account_id) do
     raise ArgumentError,
-      message: PontoConnect.token_argument_error_msg("Transactions", other)
+      message: PontoConnect.RequestUtils.token_argument_error_msg("Transactions", other)
   end
 
   @doc """
@@ -321,7 +321,7 @@ defmodule Ibanity.PontoConnect.Transaction do
   """
   def find(%Request{customer_access_token: customer_access_token} = request, ids)
       when not is_nil(customer_access_token) do
-    formatted_ids = PontoConnect.format_ids(ids)
+    formatted_ids = PontoConnect.RequestUtils.format_ids(ids)
 
     request
     |> Request.ids(formatted_ids)
@@ -336,7 +336,7 @@ defmodule Ibanity.PontoConnect.Transaction do
 
   def find(other, _ids) do
     raise ArgumentError,
-      message: PontoConnect.token_argument_error_msg("Transaction", other)
+      message: PontoConnect.RequestUtils.token_argument_error_msg("Transaction", other)
   end
 
   @doc """
@@ -381,7 +381,7 @@ defmodule Ibanity.PontoConnect.Transaction do
         ids
       )
       when not is_nil(customer_access_token) do
-    formatted_ids = PontoConnect.format_ids(ids)
+    formatted_ids = PontoConnect.RequestUtils.format_ids(ids)
 
     request
     |> Request.ids(formatted_ids)
@@ -396,7 +396,7 @@ defmodule Ibanity.PontoConnect.Transaction do
 
   def find_pending_for_account(other, _ids) do
     raise ArgumentError,
-      message: PontoConnect.token_argument_error_msg("Transaction", other)
+      message: PontoConnect.RequestUtils.token_argument_error_msg("Transaction", other)
   end
 
   @doc false
