+++ lib/ibanity/api/ponto_connect/user_info.ex
@@ -49,7 +49,7 @@ defmodule Ibanity.PontoConnect.UserInfo do
 
   def myself(other) do
     raise ArgumentError,
-      message: PontoConnect.token_argument_error_msg("UserInfo", other)
+      message: PontoConnect.RequestUtils.token_argument_error_msg("UserInfo", other)
   end
 
   @doc false
